move slave cluster to slave_sync package
the slave cluster type is only really meant for the slave sync feature,
so moving it to the slave cluster allows us to specialize it for that purpose
