+++ nbd/ardb/storage/cluster_test.go
@@ -484,26 +484,26 @@ func TestMapErrorToBroadcastStatus(t *testing.T) {
 	assert := assert.New(t)
 
 	// unknown errors return true
-	status := mapErrorToBroadcastStatus(errors.New("foo"))
+	status := MapErrorToBroadcastStatus(errors.New("foo"))
 	assert.Equal(log.StatusUnknownError, status)
 
 	// all possible sucesfull map scenarios:
 
 	// map EOF errors
-	status = mapErrorToBroadcastStatus(io.EOF)
+	status = MapErrorToBroadcastStatus(io.EOF)
 	assert.Equal(log.StatusServerDisconnect, status)
 
 	// map net.Errors
-	status = mapErrorToBroadcastStatus(stubNetError{false, false})
+	status = MapErrorToBroadcastStatus(stubNetError{false, false})
 	assert.Equal(log.StatusUnknownError, status)
 
-	status = mapErrorToBroadcastStatus(stubNetError{false, true})
+	status = MapErrorToBroadcastStatus(stubNetError{false, true})
 	assert.Equal(log.StatusServerTempError, status)
 
-	status = mapErrorToBroadcastStatus(stubNetError{true, false})
+	status = MapErrorToBroadcastStatus(stubNetError{true, false})
 	assert.Equal(log.StatusServerTimeout, status)
 
-	status = mapErrorToBroadcastStatus(stubNetError{true, true})
+	status = MapErrorToBroadcastStatus(stubNetError{true, true})
 	assert.Equal(log.StatusServerTimeout, status)
 }
